chore(xlsx): consolidate manual-layout parsers via chart/layout
Introduce src/xlsx/chart/layout.ts with the generic
parseManualLayout(parent) walker for the <c:layout><c:manualLayout>
block every chart-frame slot (chart-title, axis-title, legend, plot
area) carries to declare a custom position / size, plus the shared
readLayoutCoordinate helper that parses a single <c:x>/<c:y>/<c:w>/<c:h>
val attribute.

Replace the four duplicated walks in chart-reader.ts
(parseTitleLayout, parseLegendLayout, parsePlotAreaLayout,
parseAxisTitleLayout) with one-line delegations to parseManualLayout.
Each per-host wrapper still performs its own findChild scope check
(legend / title / axis-title) — that keeps the host-specific scope
commentary attached at the call site — and forwards to the shared
walker for the inner manualLayout coordinate extraction.

Drop the local readLayoutCoordinate definition since the layout module
now owns it. Re-export everything through the chart barrel
(src/xlsx/chart/index.ts) so future consumers can import from a
single entry point.

JSDoc preserved verbatim. All 7362 tests still pass.
